added react routes
