+++ js/components/Shop.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import ShopPage from './ShopPage';
+
+export default class extends React.Component{
+	render() {
+		return (
+  				<ShopPage data="../../api/shop.json" />
+		);
+	}
+}
