+++ package.json
@@ -4,7 +4,7 @@
   "description": "foundation app with react",
   "main": "server.js",
   "scripts": {
-    "start": "webpack-dev-server --inline --hot",
+    "start": "webpack-dev-server --inline --hot . --history-api-fallback",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
